# doc/swagger.yaml
openapi: 3.0.3
info:
  title: Sky's Auth System
  description: This is the authentication system for Sky's webservers and websites.<br>
    Some endpoints require "Super Admin" authentication. These will be marked as such, 
    and can only be used by users with the appropriate permissions. These generally
    include the project endpoints, and some administration endpoints.
  contact:
    email: auth@skystuff.cc
  version: 1.0.0
servers:
  - url: https://auth.skystuff.cc/api/
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /project/search/:
    get:
      tags:
        - Projects
      summary: Search all public projects. Unauthenticated.
      parameters:
        - in: query
          name: q
          schema:
            type: string
          description: The case insensitive project name to search for.
        - in: query
          name: limit
          schema:
            type: integer
          description: The total number of projects to list. Maximum of 50.
        - in: query
          name: offset
          schema:
            type: integer
          description: Offset for search results.
      description: Searches and returns a list of projects based on the query. Pass no query to list all projects. If you are looking for projects the user has access to, use [/user/projects/](#/User/user/projects/).
      operationId: project.search_projects
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Projects'
  /project/details/{project_name}/:
    get:
      tags:
        - Projects
      summary: Get details about a specific project. Unauthenticated for public projects.
      parameters:
        - in: path
          name: project_name
          schema:
            type: string
          description: The case insensitive project name to search for.
          required: true
      operationId: project.project_details
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/ProjectDetails'
  /project/enroll/:
    post:
      tags:
        - Projects
      summary: Apply to be added to a project.
      description: Apply to be added to a project. If the project is public, user will be added automatically. If it is a private project, the user's status will change to 'pending', but will still be treated as denied. If the user is explicitly denied, they can not re-apply. If the user is allowed, their status will be 'approved'
      operationId: project.enroll
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/project.enroll'
      responses:
        '204':
          description: Successful operation.
        '400':
          description: User already applied.
        '401':
          description: Unauthorized.
        '403':
          description: User is denied from project. Contact admin to grant access.
  /project/status/{project_name}/:
    get:
      tags:
        - Projects
      summary: Get the users status for a project.
      description: Gets the users status for a project. This will return a ProjectStatus json.
      parameters:
        - in: path
          name: project_name
          schema:
            type: string
          required: true
          description: Case insensitive project name to get status of.
      operationId: project.get_status
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectStatus'
        '401':
          description: Unauthorized.
        '404':
          description: Project not found.
  
  # these are admin endpoints, user must be authenticated
  /project/create/:
    post:
      tags:
        - Projects
      summary: Create a project. Requires Super Admin.
      operationId: project.create_project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
      responses:
        '204':
          description: Successful operation.
        '400':
          description: Missing body element.
        '401':
          description: Unuathorized.
  /project/edit/{project_name}/:
    patch:
      tags:
        - Projects
      summary: Edit a project. Requires Super Admin.
      operationId: project.edit_project
      parameters:
        - in: path
          name: project_name
          schema:
            type: string
          required: true
          description: Case insensitive project name to edit.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
      responses:
        '204':
          description: Successful operation.
        '401':
          description: Unauthorized.
        '404':
          description: Project not found.
...
  /project/getpending/{project_name}/:
    get:
      tags:
        - Projects
      summary: Get users awaiting approval for projects. Requires Super Admin.
      description: Get pending approvals for projects. Requires Super Admin to get the list, and requries Super Admin to POST approval back to server.
      parameters:
        - in: path
          name: project_name
          schema:
            type: string
          required: true
        - in: query
          name: limit
          schema:
            type: integer
          description: The total number of users to list. Maximum of 50.
        - in: query
          name: offset
          schema:
            type: integer
          description: Offset for search results.
      operationId: project.get_pending_approvals
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AwaitingApprovals'
        '401':
          description: Unauthorized.
  /project/approve/:
    post:
      tags:
        - Projects
      summary: Approve/Deny a user for a project. Requires Super Admin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/project.approval'
      operationId: project.approve
      responses:
        '204':
          description: Successful operation.
        '400':
          description: Missing body field.
        '401':
          description: Unauthorized.
        '404':
          description: User not found. Project not found.
  /project/getusers/{project_name}/:
    get:
      tags:
        - Projects
      summary: Get users of a project. Requires Super Admin.
      description: Gets a list of users for a project. Will also return users that are pending approval.
      operationId: project.get_project_users
      parameters:
        - in: path
          name: project_name
          schema:
            type: string
          required: true
          description: Case insensitive project name to get.
        - in: query
          name: limit
          schema:
            type: integer
          description: The total number of users to list. Maximum of 50.
        - in: query
          name: offset
          schema:
            type: integer
          description: Offset for search results.
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectUsers'
        '401':
          description: Unauthorized
components:
  schemas:
...
    PartialUser:
      type: object
      properties:
        name:
          type: string
          example: Sky
        super_admin:
          type: boolean
          description: If this field is set, projects should treat the user as an admin.
          example: true
    Project:
      type: object
      properties:
        name:
          type: string
          example: paste
        public:
          type: boolean
          description: This boolean is whether or not anybody can see this project. If it is false, users will be specifically added to it by an admin.
          example: false
        open:
          type: boolean
          description: This boolean is whether or not anybody can join this project. If it is false, users must be approved for the project.
          example: true
        description:
          type: string
          description: This is a generic description of the project. Markdown formatted.
    Projects:
      type: array
      items:
        $ref: '#/components/schemas/Project'
    ProjectStatus:
      allOf:
        - $ref: '#/components/schemas/Project'
        - type: object
          properties:
            approval:
              type: string
              description: The user's approval status for a project. One of; "approved", "denied", "pending", "inactive".
              example: inactive
...
    AwaitingApproval:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/PartialUser'
        project:
          $ref: '#/components/schemas/Project'
    AwaitingApprovals:
      type: array
      items:
        $ref: '#/components/schemas/AwaitingApproval'
    ProjectUser:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/PartialUser'
        approval:
          type: string
          description: The user's approval status for a project. One of; "approved", "pending".
          example: pending
    ProjectUsers:
      type: object
      allOf:
        - $ref: '#/components/schemas/Project'
        - type: object
          properties:
            users:
              type: array
              items:
                $ref: '#/components/schemas/ProjectUser'
    ProjectDetail:
      type: object
      allOf:
        - $ref: '#/components/schemas/Project'
        - type: object
          properties:
            user_count:
              type: integer
    ProjectDetails:
      type: array
      items:
        $ref: '#/components/schemas/ProjectDetail'
...
    project.enroll:
      type: object
      properties:
        name:
          type: string
          description: Name of the project to enroll for.
          example: paste
    project.approval:
      type: object
      properties:
        user:
          type: string
          description: Name of the user that is being approved/denied.
        project:
          type: string
          description: Name of project that is being approved/denied
        approved:
          type: boolean
          description: Approval status. True if approved, false if denied.
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: Authorization